major updates in structure of signal classes, matrix param obj

+++ src/ParamObjects/JPetLayer/JPetLayerFactory.cpp
@@ -1,5 +1,5 @@
 /**
- *  @copyright Copyright 2019 The J-PET Framework Authors. All rights reserved.
+ *  @copyright Copyright 2021 The J-PET Framework Authors. All rights reserved.
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may find a copy of the License in the LICENCE file.
@@ -21,39 +21,44 @@
 
 std::map<int, JPetLayer*>& JPetLayerFactory::getLayers()
 {
-  if (!fInitialized) { initialize(); }
+  if (!fInitialized)
+  {
+    initialize();
+  }
   return fLayers;
 }
 
 void JPetLayerFactory::initialize()
 {
-  ParamObjectsDescriptions descriptions = fParamGetter.getAllBasicData(
-    ParamObjectType::kLayer, fRunID
-  );
-  if (descriptions.size() == 0) {
+  ParamObjectsDescriptions descriptions = fParamGetter.getAllBasicData(ParamObjectType::kLayer, fRunID);
+  if (descriptions.size() == 0)
+  {
     ERROR(Form("No layers in run %i", fRunID));
     return;
   }
-  for (auto description : descriptions) {
+  for (auto description : descriptions)
+  {
     fLayers[description.first] = build(description.second);
   }
   fInitialized = true;
-  ParamRelationalData relations = fParamGetter.getAllRelationalData(
-    ParamObjectType::kLayer, ParamObjectType::kSetup, fRunID
-  );
-  for (auto relation : relations) {
+  ParamRelationalData relations = fParamGetter.getAllRelationalData(ParamObjectType::kLayer, ParamObjectType::kSetup, fRunID);
+  for (auto relation : relations)
+  {
     fLayers[relation.first]->setSetup(*fSetupFactory.getSetups().at(relation.second));
   }
 }
 
 JPetLayer* JPetLayerFactory::build(ParamObjectDescription data)
 {
-  try {
+  try
+  {
     int id = boost::lexical_cast<int>(data.at("id"));
     std::string name = data.at("name");
-    float radius = boost::lexical_cast<float>(data.at("radius"));
+    double radius = boost::lexical_cast<double>(data.at("radius"));
     return new JPetLayer(id, name, radius);
-  } catch (const std::exception & e) {
+  }
+  catch (const std::exception& e)
+  {
     ERROR(Form("Failed to build layer with error: %s", e.what()));
     throw;
   }
